move stats into subpackage
