+++ src/lenskit/stats/__init__.py
@@ -1,0 +1,13 @@
+# This file is part of LensKit.
+# Copyright (C) 2018-2023 Boise State University.
+# Copyright (C) 2023-2025 Drexel University.
+# Licensed under the MIT license, see LICENSE.md for details.
+# SPDX-License-Identifier: MIT
+
+from ._gini import gini
+from ._topn import argtopn
+
+__all__ = [
+    "gini",
+    "argtopn",
+]
