+++ src/lenskit/stats/_gini.py
@@ -4,10 +4,6 @@
 # Licensed under the MIT license, see LICENSE.md for details.
 # SPDX-License-Identifier: MIT
 
-"""
-LensKit statistical computations.
-"""
-
 from __future__ import annotations
 
 import warnings
@@ -15,7 +11,6 @@ import warnings
 import numpy as np
 from numpy.typing import ArrayLike
 
-from lenskit.data.types import NPVector
 from lenskit.diagnostics import DataWarning
 
 
@@ -59,40 +54,4 @@ def gini(xs: ArrayLike) -> float:
         warnings.warn(
             "Gini coefficient is not defined for non-positive totals", DataWarning, stacklevel=2
         )
-    return max(num / denom, 0)
-
-
-def argtopn(xs: ArrayLike, n: int) -> NPVector[np.int64]:
-    """
-    Compute the ordered positions of the top *n* elements.  Similar to
-    :func:`torch.topk`, but works with NumPy arrays and only returns the
-    indices.
-
-    .. deprecated:: 2025.3.0
-
-        This was never declared stable, but is now deprecated and will be
-        removed in 2026.1.
-    """
-    if n == 0:
-        return np.empty(0, np.int64)
-
-    xs = np.asarray(xs)
-
-    N = len(xs)
-    invalid = np.isnan(xs)
-    if np.any(invalid):
-        mask = ~invalid
-        vxs = xs[mask]
-        remap = np.arange(N)[mask]
-        res = argtopn(vxs, n)
-        return remap[res]
-
-    if n >= 0 and n < N:
-        parts = np.argpartition(-xs, n)
-        top_scores = xs[parts[:n]]
-        top_sort = np.argsort(-top_scores)
-        order = parts[top_sort]
-    else:
-        order = np.argsort(-xs)
-
-    return order
+    return max(num / denom, 0.0)
